Move Observable class
